requirements.txt
pytest
matplotlib
scikit-learn
scipy
spacy
numpy
einops
einsum
fvcore
h5py
twine
transformers==4.45.2
huggingface_hub==0.24
accelerate==0.34.2
diffusers==0.30.3
pillow==10.2.0
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
flash-attn==2.6.3
git+https://github.com/Dao-AILab/flash-attention@v2.6.3#subdirectory=csrc/fused_dense_lib
jaxtyping
mediapipe
gradio
git+https://github.com/facebookresearch/pytorch3d.git
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
